first move of precomp to its own subdirectory- maybe like the old days
