+++ wisdem/precomp/__init__.py
@@ -1,0 +1,5 @@
+from wisdem.precomp.precomp import PreComp
+from wisdem.precomp.profile import Profile
+from wisdem.precomp.composite_section import CompositeSection
+from wisdem.precomp.orthotropic import Orthotropic2DMaterial
+
