+++ wisdem/rotorse/rotor_elasticity.py
@@ -4,7 +4,7 @@ import numpy as np
 from openmdao.api import Group, ExplicitComponent
 from scipy.interpolate import PchipInterpolator
 
-from wisdem.rotorse.precomp import PreComp, Profile, CompositeSection, Orthotropic2DMaterial
+from wisdem.precomp import PreComp, Profile, CompositeSection, Orthotropic2DMaterial
 from wisdem.commonse.utilities import rotate, arc_length
 from wisdem.rotorse.rail_transport import RailTransport
 import logging
